backend/requirements.txt
fastapi
uvicorn[standard]
httpx
google-generativeai
python-dotenv
pydantic
